add detail/parsing/binary_reader.hpp
